+++ src/qmflows/data/dictionaries/propertiesDFTB.yaml
@@ -4,46 +4,46 @@ charges:
     kwargs:
         section: AMSResults
         prop: Charges
-    parser: adf_parser
+    parser: adf
 dipole:
     file_ext: rkf
-    function: kfreader    
+    function: kfreader
     kwargs:
         section: AMSResults
         prop: DipoleMoment
-    parser: adf_parser
+    parser: adf
 energy:
     file_ext: rkf
     function: kfreader
     kwargs:
         section: AMSResults
         prop: Energy
-    parser: adf_parser
+    parser: adf
 enthalpy:
     file_ext: rkf
     function: kfreader
     kwargs:
         section: Thermodynamics
         prop: Enthalpy
-    parser: adf_parser
+    parser: adf
 free_energy:
     file_ext: rkf
     function: kfreader
     kwargs:
         section: Thermodynamics
         prop: 'Gibbs free Energy'
-    parser: adf_parser
+    parser: adf
 frequencies:
     file_ext: rkf
     function: kfreader
     kwargs:
         prop: Frequencies[cm-1]
         section: Vibrations
-    parser: adf_parser
+    parser: adf
 hessian:
     file_ext: rkf
     function: kfreader
     kwargs:
         section: AMSResults
         prop: Hessian
-    parser: adf_parser
+    parser: adf
