+++ src/qmflows/parsers/__init__.py
@@ -1,8 +1,23 @@
-"""Parsers API."""
+"""Parsers API.
 
-from .generic_parsers import (awk_file, extract_line_value, extract_line_values)
-from .xyzParser import (parse_string_xyz, readXYZ,
-                        manyXYZ, string_to_plams_Molecule)
+Consists of the following four public submodules:
+
+.. automodule:: qmflows.parsers.adf
+.. automodule:: qmflows.parsers.cp2k
+.. automodule:: qmflows.parsers.orca
+.. automodule:: qmflows.parsers.utils
+
+"""
+
+from ._generic import (awk_file, extract_line_value, extract_line_values)
+from ._xyz import (parse_string_xyz, readXYZ, manyXYZ, string_to_plams_Molecule)
+
+from . import (
+    adf as adf,
+    cp2k as cp2k,
+    orca as orca,
+    utils as utils,
+)
 
 __all__ = [
     'awk_file', 'extract_line_value', 'extract_line_values', 'parse_string_xyz',
